Empty file header
The default file header contains a little Javadoc block with author
information, without using the @author tag. Even if it did use the tag,
@author is a lie waiting to happen and adds no -- and sometimes negative
-- value.

This commit empties the file header so that it produces no output when
rendered. This is to facilitate a file header for those that really need
it while avoiding having to remove the instruction from every template
that defines it.