Enforce singleton property with enum
Per Effective Java, 2nd Edition, Item 3, this commit changes the default
singleton template to use an enum instead of a class. Doing so
automatically solves all serialization issues. However, because 1) this
pattern remains rare in the wild (even for singletons, which should be
rare on their own); 2) it unavoidably clutters the interface with enum
related methods; and 3) enum constructors are always private, the
classic implementation is moved to another file.